requirements.txt
Scrapy==1.8.0
mysql-connector-python==8.0.18
pandas==0.25.0
Pillow==5.4.1
selenium==3.141.0
pytesseract==0.2.6
wget==3.2
xlrd==1.2.0
apache-airflow[mysql]==1.10.6
werkzeug==0.16.0
flask_bcrypt
